+++ test/deterministic_sampler_test.exs
@@ -1,0 +1,4 @@
+defmodule HoneylixirDeterminsticSamplerTest do
+  use ExUnit.Case, async: true
+  doctest Honeylixir.DeterminsticSampler
+end